Add after and before dates parameters for getting ers messages from api
