+++ backend/src/main/kotlin/fr/gouv/cnsp/monitorfish/domain/repositories/ERSRepository.kt
@@ -9,8 +9,9 @@ import kotlin.jvm.Throws
 interface ERSRepository {
     @Throws(NoERSLastDepartureDateFound::class)
     fun findLastDepartureDateAndTripNumber(internalReferenceNumber: String): LastDepartureDateAndTripNumber
-    fun findAllMessagesAfterDepartureDate(dateTime: ZonedDateTime,
-                                          internalReferenceNumber: String): List<ERSMessage>
+    fun findAllMessagesBetweenDepartureDates(afterDateTime: ZonedDateTime,
+                                             beforeDateTime: ZonedDateTime,
+                                             internalReferenceNumber: String): List<ERSMessage>
     fun findLANAndPNOMessagesNotAnalyzedBy(ruleType: String): List<Pair<ERSMessage, ERSMessage?>>
     fun updateERSMessagesAsProcessedByRule(ids: List<Long>, ruleType: String)
     fun findById(id: Long): ERSMessage
